refactor: refactor contentEditable handlers into one file
