+++ spec/ContentEditableDivSpec.test.js
@@ -1,4 +1,7 @@
-import { handleContentEditableChange } from '../src/callbacks/ContentEditableChange';
+import {
+  handleKeyPress,
+  handleCarriageReturn,
+} from '../src/callbacks/ContentEditableDiv.js';
 
 const buildMutation = (type, target) => {
   return { type, target };
@@ -7,7 +10,7 @@ const buildMutation = (type, target) => {
 describe('handleContentEditableDivChange', () => {
   const action = (node) => console.log(node);
   let spy;
-  let handler = handleContentEditableChange(action);
+  let handler = handleKeyPress(action);
 
   beforeEach(() => {
     spy = jest.spyOn(console, 'log').mockImplementationOnce(() => {});
@@ -51,6 +54,41 @@ describe('handleContentEditableDivChange', () => {
       handler([mutation]);
 
       expect(spy).not.toHaveBeenCalledWith(mutation);
+    });
+  });
+
+  describe('#handleCarriageReturn', () => {
+    let mockNode;
+    beforeEach(() => {
+      mockNode = {
+        dispatchEvent: jest.fn(),
+        hasAttribute: () => true,
+      };
+    });
+    describe('when an observed content editable sees a non-text change (such as a line break)', () => {
+      it('should dispatch an input event to trigger checking for warnings', () => {
+        const mockMutation = {
+          type: 'childList',
+          target: mockNode,
+        };
+
+        handleCarriageReturn([mockMutation]);
+
+        expect(mockNode.dispatchEvent).toHaveBeenCalledWith(expect.any(Event));
+      });
+    });
+
+    describe('when an observed content editable sees a text change', () => {
+      it('should NOT dispatch an extra input event', () => {
+        const mockMutation = {
+          type: 'characterData',
+          target: mockNode,
+        };
+
+        handleCarriageReturn([mockMutation]);
+
+        expect(mockNode.dispatchEvent).not.toHaveBeenCalled();
+      });
     });
   });
 });
