--- src/callbacks/ContentEditableChange.js
@@ -1,6 +1,0 @@
-const isContentEditableChildList = (mutation) =>
-  mutation.type === 'childList' &&
-  mutation.target.hasAttribute('contentEditable');
-
-export const handleContentEditableChange = (action) => (mutations) =>
-  mutations.filter(isContentEditableChildList).forEach(action);
